패치로직 수정
