+++ src/domains/community/main/PostCard.tsx
@@ -14,22 +14,11 @@ import CommunitySkeleton from './CommunitySkeleton';
 type Props = {
   posts: Post[];
   setPosts: React.Dispatch<React.SetStateAction<Post[]>>;
+  isLoading: boolean;
 };
 
-function PostCard({ posts, setPosts }: Props) {
+function PostCard({ posts, setPosts, isLoading }: Props) {
   const router = useRouter();
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      setIsLoading(true);
-      const data = await fetchPost();
-      if (!data) return;
-      setPosts(data);
-      setIsLoading(false);
-    };
-    fetchData();
-  }, [setPosts]);
 
   const handlePost = (id: number) => {
     router.push(`/community/${id}`);
@@ -43,64 +32,51 @@ function PostCard({ posts, setPosts }: Props) {
 
   return (
     <>
-      {posts.map(
-        ({
-          postId,
-          categoryName,
-          title,
-          content,
-          userNickName,
-          viewCount,
-          createdAt,
-          imageUrl,
-        }) => (
-          <article className="pt-5 pb-3 border-b-1 border-gray-light" key={postId}>
-            <Label title={categoryName} />
+      {posts.map(({ postId, categoryName, title, content, userNickName, viewCount, createdAt }) => (
+        <article className="pt-5 pb-3 border-b-1 border-gray-light" key={postId}>
+          <Label title={categoryName} />
 
-            <section
-              onClick={() => handlePost(postId)}
-              className="flex items-start justify-between mt-3 cursor-pointer h-full content-between"
-              role="link"
-            >
-              <div className="flex flex-col gap-2 max-w-[860px] flex-grow">
-                <p className="font-bold sm:text-xl text-lg">
-                  일이삼사오육칠팔구십일이삼사오육칠팔구십
+          <section
+            onClick={() => handlePost(postId)}
+            className="flex items-start justify-between mt-3 cursor-pointer h-full content-between"
+            role="link"
+          >
+            <div className="flex flex-col gap-2 max-w-[860px] flex-grow">
+              <p className="font-bold sm:text-xl text-lg">{title}</p>
+              <div className="font-light sm:text-[15px] text-sm md:max-w-[850px] sm:max-w-[500px] max-w-[210px] h-full">
+                <p
+                  className="h-10"
+                  style={{
+                    display: '-webkit-box',
+                    WebkitBoxOrient: 'vertical',
+                    WebkitLineClamp: 2,
+                    overflow: 'hidden',
+                    textOverflow: 'ellipsis',
+                  }}
+                >
+                  {content}
                 </p>
-                <div className="font-light sm:text-[15px] text-sm md:max-w-[850px] sm:max-w-[500px] max-w-[210px] h-full">
-                  <p
-                    className="h-10"
-                    style={{
-                      display: '-webkit-box',
-                      WebkitBoxOrient: 'vertical',
-                      WebkitLineClamp: 2,
-                      overflow: 'hidden',
-                      textOverflow: 'ellipsis',
-                    }}
-                  >
-                    {content}
-                  </p>
-                </div>
-                <PostInfo
-                  hasUserName={true}
-                  userNickName={userNickName}
-                  viewCount={viewCount}
-                  createdAt={createdAt}
-                  postId={postId}
-                />
               </div>
-              <figure className="flex items-center flex-shrink-0 md:w-[115px] md:h-[115px] w-[85px] h-[85px]">
-                <Image
-                  src={prePost}
-                  alt="예비사진"
-                  width={105}
-                  height={105}
-                  className="w-full h-full object-cover self-center"
-                />
-              </figure>
-            </section>
-          </article>
-        )
-      )}
+              <PostInfo
+                hasUserName={true}
+                userNickName={userNickName}
+                viewCount={viewCount}
+                createdAt={createdAt}
+                postId={postId}
+              />
+            </div>
+            <figure className="flex items-center flex-shrink-0 md:w-[115px] md:h-[115px] w-[85px] h-[85px]">
+              <Image
+                src={prePost}
+                alt="예비사진"
+                width={105}
+                height={105}
+                className="w-full h-full object-cover self-center"
+              />
+            </figure>
+          </section>
+        </article>
+      ))}
     </>
   );
 }
